wip: move config package to dotenv
