+++ .gitignore
@@ -44,3 +44,5 @@ dist
 !.env.template
 
 .turbo
+
+out/
