+++ packages/dotenv/package.json
@@ -1,5 +1,5 @@
 {
-    "name": "@qnaplus/config",
+    "name": "@qnaplus/dotenv",
     "version": "1.0.0",
     "description": "",
     "main": "dist/index.js",
