+++ packages/qnaplus/package.json
@@ -14,7 +14,7 @@
         "yarn": "1.22.22"
     },
     "dependencies": {
-        "@qnaplus/store": "1.0.0",
-        "@qnaplus/config": "1.0.0"
+        "@qnaplus/store": "*",
+        "@qnaplus/dotenv": "*"
     }
 }
