+++ packages/qnaplus/src/index.ts
@@ -1,2 +1,2 @@
-export * from "@qnaplus/config";
+export * from "@qnaplus/dotenv";
 export * from "@qnaplus/store";
