+++ turbo.json
@@ -2,8 +2,8 @@
     "$schema": "https://turbo.build/schema.json",
     "tasks": {
       "build": {
-        "outputs": ["dist/**"],
-        "dependsOn": ["^build"]
+        "dependsOn": ["^build"],
+        "outputs": ["dist/**"]
       },
       "check-types": {
         "dependsOn": ["^check-types"]
